* master-fix: (144 commits)
  PROV-1395 Change behavior so storage locations are evaluated based upon starting date only, and assumed to run until overridden by a later dated location record. Most (ok, all) users are logging storage locations using a date of placement, not the range of dates the item is present. The current implementation assumes a range.
  PROV-1529 Allow BCE start dates
  PROV-1528 remove debugging alert
  PROV-1527 Use correct values when sorting on authority attributes, avoiding SQL error
  PROV-1525 Catch empty plugin
  NOJIRA Identify AVC as video/mp4. video/quicktime is being run through the QuicktimeVR plugin by default.
  PROV-1524 Pull effective_date and source_info from interstitialValues option so they get passed to addRelationship() and are not processed after the fact
  NOJIRA Add GRAM profile
  PROV-1469 Add matching fix
  NOJIRA Re-add code lost in merge
  PROV-1523 implement new on-load behavior
  NOJIRA Simplify relationship bundle preview code
  PROV-1519 More reasonable test case
  PROV-1521 Pass delimiter option
  PROV-1519 Fix issue that could lead to attributes being treated as intrinsics in ElasticSearch query rewrite
  PROV-1519 Allow setting max_result_window
  PROV-1519 Pull in change from https://github.com/elastic/elasticsearch-php/pull/364 that hasn't made it into a release yet
  PROV-1519 Streamline install process
  Bump version
  PROV-1519 Fix potential exception when running the installer with ElasticSearch
  ...

+++ app/lib/core/Plugins/SearchEngine/ElasticSearch/FieldTypes/Numeric.php
@@ -34,7 +34,7 @@ namespace ElasticSearch\FieldTypes;
 
 require_once(__CA_LIB_DIR__.'/core/Plugins/SearchEngine/ElasticSearch/FieldTypes/GenericElement.php');
 
-class Float extends GenericElement {
+class Numeric extends GenericElement {
 
 	public function __construct($ps_table_name, $ps_element_code) {
 		parent::__construct($ps_table_name, $ps_element_code);
